NTT-18 - Addressed review comments
