+++ app/controllers/SessionExpiredController.scala
@@ -25,9 +25,9 @@ import uk.gov.hmrc.play.bootstrap.controller.FrontendBaseController
 import scala.concurrent.ExecutionContext
 
 class SessionExpiredController @Inject()(
-    val controllerComponents: MessagesControllerComponents,
-    renderer: Renderer
-)(implicit ec: ExecutionContext) extends FrontendBaseController with I18nSupport {
+                                          val controllerComponents: MessagesControllerComponents,
+                                          renderer: Renderer
+                                        )(implicit ec: ExecutionContext) extends FrontendBaseController with I18nSupport {
 
   def onPageLoad: Action[AnyContent] = Action.async {
     implicit request =>
