--- app/models/TrusteesBasedInUK.scala
@@ -1,51 +1,0 @@
-/*
- * Copyright 2020 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package models
-
-import play.api.data.Form
-import play.api.i18n.Messages
-import uk.gov.hmrc.viewmodels._
-
-sealed trait TrusteesBasedInUK
-
-object TrusteesBasedInUK extends Enumerable.Implicits {
-
-  case object All extends WithName("all") with TrusteesBasedInUK
-  case object None extends WithName("none") with TrusteesBasedInUK
-  case object Both extends WithName("both") with TrusteesBasedInUK
-
-  val values: Seq[TrusteesBasedInUK] = Seq(
-    All,
-    None,
-    Both
-  )
-
-  def checkboxes(form: Form[_])(implicit messages: Messages): Seq[Checkboxes.Item] = {
-
-    val field = form("value")
-    val items = Seq(
-      Checkboxes.Checkbox(msg"trusteesBasedInUK.all", All.toString),
-      Checkboxes.Checkbox(msg"trusteesBasedInUK.none", None.toString),
-      Checkboxes.Checkbox(msg"trusteesBasedInUK.both", Both.toString)
-    )
-
-    Checkboxes.set(field, items)
-  }
-
-  implicit val enumerable: Enumerable[TrusteesBasedInUK] =
-    Enumerable(values.map(v => v.toString -> v): _*)
-}
