--- conf/views/trusteesBasedInUK.njk
@@ -1,57 +1,0 @@
-{% extends "includes/layout.njk" %}
-
-{% from "govuk/components/back-link/macro.njk"     import govukBackLink %}
-{% from "govuk/components/checkboxes/macro.njk"    import govukCheckboxes %}
-{% from "govuk/components/button/macro.njk"        import govukButton %}
-{% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
-{% from "macros/title.njk"                         import title %}
-
-{% block pageTitle %}
-  {{ title(messages("trusteesBasedInUK.title"), form.errors) }}
-{% endblock %}
-
-{% block mainContent %}
-
-  {{ govukBackLink({
-    text: messages("site.back")
-  }) }}
-
-  <div class="govuk-main-wrapper">
-    <div class="govuk-grid-row">
-      <div class="govuk-grid-column-two-thirds">
-
-        {% if form.errors.length > 0 %}
-          {{ govukErrorSummary({
-            "titleText": messages("error.summary.title"),
-            "errorList": form.errors
-          }) }}
-        {% endif %}
-
-        <form action="{{ routes.controllers.TrusteesBasedInUKController.onSubmit(mode).url }}" method="post">
-
-          {{ csrf() | safe }}
-
-          {{ govukCheckboxes({
-            name: 'value',
-            fieldset: {
-              legend: {
-                text: messages('trusteesBasedInUK.title'),
-                isPageHeading: true,
-                classes: "govuk-fieldset__legend--xl"
-              }
-            },
-            items: checkboxes,
-            errorMessage: form.value.error
-          }) }}
-
-          {{ govukButton({
-            text: messages("site.continue")
-          }) }}
-
-        </form>
-
-      </div>
-    </div>
-  </div>
-
-{% endblock %}
