+++ test/generators/UserAnswersGenerator.scala
@@ -27,6 +27,7 @@ trait UserAnswersGenerator extends TryValues {
   self: Generators =>
 
   val generators: Seq[Gen[(QuestionPage[_], JsValue)]] =
+    arbitrary[(TrusteesBasedInUKPage.type, JsValue)] ::
     arbitrary[(PreviouslyBasedPage.type, JsValue)] ::
     arbitrary[(OffshoreTrustPage.type, JsValue)] ::
     arbitrary[(ExpressTrustPage.type, JsValue)] ::
